+++ src/sentry/hybridcloud/tasks/__init__.py
@@ -1,1 +1,3 @@
+from .backfill_outboxes import *  # noqa
+from .deliver_from_outbox import *  # noqa
 from .deliver_webhooks import *  # noqa
