+++ src/sentry/testutils/outbox.py
@@ -8,9 +8,12 @@ from django.conf import settings
 from django.core.handlers.wsgi import WSGIRequest
 
 from sentry.hybridcloud.models.webhookpayload import THE_PAST, WebhookPayload
+from sentry.hybridcloud.tasks.deliver_from_outbox import (
+    enqueue_outbox_jobs,
+    enqueue_outbox_jobs_control,
+)
 from sentry.models.outbox import OutboxBase
 from sentry.silo.base import SiloMode
-from sentry.tasks.deliver_from_outbox import enqueue_outbox_jobs, enqueue_outbox_jobs_control
 from sentry.testutils.silo import assume_test_silo_mode
